requirements.txt
# requirements.txt
flask==2.0.1
werkzeug==2.0.1
python-dotenv==0.21.1
pytube==12.1.2
openai==0.27.4
requests==2.28.2
uuid==1.30
